Switched from float to double.

+++ src/map/src/scan_match.h
@@ -9,26 +9,26 @@
 struct icpConfig {
 	int minICPComparePoints;
 	int maxICPLoopCount;
-	float icpAverageDistanceTraveledThresholdSquared;
+	double icpAverageDistanceTraveledThresholdSquared;
 	int icpNoMovementCounterThreshold;
-	float goodCorrespondenceThresholdSquared;
-	float maximumPointMatchDistance;
+	double goodCorrespondenceThresholdSquared;
+	double maximumPointMatchDistance;
 };
 struct icpOutput {
-	std::vector<std::vector<float>> rotationMatrix;
-	std::vector<float> translation;
-	float theta;
+	std::vector<std::vector<double>> rotationMatrix;
+	std::vector<double> translation;
+	double theta;
 };
 struct svdOutput {
-	std::vector<std::vector<float>> U;
-	std::vector<std::vector<float>> S;
-	std::vector<std::vector<float>> V;
+	std::vector<std::vector<double>> U;
+	std::vector<std::vector<double>> S;
+	std::vector<std::vector<double>> V;
 };
 
-std::vector<std::vector<int>> matchPoints(std::vector<std::vector<float>> pc1, std::vector<std::vector<float>> pc2);
-svdOutput SVD(std::vector<std::vector<float>> A);
-void transpose(std::vector<std::vector<float>> &m);
-icpOutput runICP(std::vector<std::vector<float>> set1, std::vector<std::vector<float>> set2);
+std::vector<std::vector<int>> matchPoints(std::vector<std::vector<double>> pc1, std::vector<std::vector<double>> pc2);
+svdOutput SVD(std::vector<std::vector<double>> A);
+void transpose(std::vector<std::vector<double>> &m);
+icpOutput runICP(std::vector<std::vector<double>> set1, std::vector<std::vector<double>> set2);
 int optimizeScan(worldState &newScan, std::vector<worldState> map, icpConfig cfg);
 
 #endif
