+++ src/map/src/world_state.cpp
@@ -13,19 +13,19 @@ worldState::worldState() {
 
 	worldState::convertToRobotFrame();
 }
-worldState::worldState(geometry_msgs::Pose odom, sensor_msgs::LaserScan base, float lidarDistance) {
+worldState::worldState(geometry_msgs::Pose odom, sensor_msgs::LaserScan base, double lidarDistance) {
 	odometry = odom;
 	baseScan = base;
 	lidarForwardDistance = lidarDistance;
 
-	theta = static_cast<float>(atan2(2*((odometry.orientation.x*odometry.orientation.y) + (odometry.orientation.z*odometry.orientation.w)), 1-(2*((odometry.orientation.y*odometry.orientation.y) + (odometry.orientation.z*odometry.orientation.z)))));
+	theta = static_cast<double>(atan2(2*((odometry.orientation.x*odometry.orientation.y) + (odometry.orientation.z*odometry.orientation.w)), 1-(2*((odometry.orientation.y*odometry.orientation.y) + (odometry.orientation.z*odometry.orientation.z)))));
 
 	worldState::convertToRobotFrame();
 	worldState::convertToWorldFrame();
 }
 
 void worldState::convertToRobotFrame() {
-	float t, tMin, tInc;
+	double t, tMin, tInc;
 	tMin = baseScan.angle_min;
 	tInc = baseScan.angle_increment;
 	t = tMin;
@@ -38,7 +38,7 @@ void worldState::convertToRobotFrame() {
 		walls[i][1] = 0;
 	}
 }
-float worldState::convertToWorldFrame() {
+double worldState::convertToWorldFrame() {
 	for(int i=0; i<sizeof(walls)/sizeof(walls[0]); ++i) {
 		walls[i][0] = (walls[i][0] * cos(theta)) - (walls[i][1] * sin(theta)) + odometry.position.x;
 		walls[i][1] = (walls[i][0] * sin(theta)) + (walls[i][1] * cos(theta)) + odometry.position.y;
@@ -47,16 +47,16 @@ float worldState::convertToWorldFrame() {
 }
 void worldState::newOdometry(geometry_msgs::Pose odom) {
 	odometry = odom;
-	theta = static_cast<float>(atan2(2*((odometry.orientation.x*odometry.orientation.y) + (odometry.orientation.z*odometry.orientation.w)), 1-(2*((odometry.orientation.y*odometry.orientation.y) + (odometry.orientation.z*odometry.orientation.z)))));
+	theta = static_cast<double>(atan2(2*((odometry.orientation.x*odometry.orientation.y) + (odometry.orientation.z*odometry.orientation.w)), 1-(2*((odometry.orientation.y*odometry.orientation.y) + (odometry.orientation.z*odometry.orientation.z)))));
 	worldState::convertToRobotFrame();
 }
-float worldState::newBaseScan(sensor_msgs::LaserScan base) {
+double worldState::newBaseScan(sensor_msgs::LaserScan base) {
 	baseScan = base;
 	worldState::convertToRobotFrame();
-	float a = worldState::convertToWorldFrame();
+	double a = worldState::convertToWorldFrame();
 	return a;
 }
-void worldState::setLidarForwardDistance(float d) {
+void worldState::setLidarForwardDistance(double d) {
 	//
 	lidarForwardDistance = d;
 }
@@ -68,20 +68,20 @@ sensor_msgs::LaserScan worldState::getBaseScan() {
 	//
 	return baseScan;
 }
-void worldState::getWalls(float (&copyIntoThis)[BASE_SCAN_MAX_NUM_POINTS][2]) {
+void worldState::getWalls(double (&copyIntoThis)[BASE_SCAN_MAX_NUM_POINTS][2]) {
 	for(int i=0; i<BASE_SCAN_MAX_NUM_POINTS; ++i) {
 		for(int j=0; j<2; ++j) {
 			copyIntoThis[i][j] = walls[i][j];
 		}
 	}
 }
-float worldState::getTheta() {
+double worldState::getTheta() {
 	//
 	return theta;
 }
 char* worldState::makeJSONString() {
-	const std::vector<float> position = {odometry.position.x, odometry.position.y, odometry.position.z};
-	const std::vector<float> orientation  = {odometry.orientation.x, odometry.orientation.y, odometry.orientation.z, odometry.orientation.w};
+	const std::vector<double> position = {odometry.position.x, odometry.position.y, odometry.position.z};
+	const std::vector<double> orientation  = {odometry.orientation.x, odometry.orientation.y, odometry.orientation.z, odometry.orientation.w};
 
 	json pose = {
 		{"position", position},
